Add metadata to main data object
There are times when templates might want access to the tags data and
not be one of the tag output files. Currently the tags data is only
available in the output file (`topics/{tag}.html`) for which your
template can iterate over the tags object.

With this feature the user can iterate over tags outside of the
output file by accessing the `metalsmith.data.tags` property (or in the
case of a template `this.tags`).