Remove build files
This is already in the gitignore file but somehow old copies were still
stored in the history. This cause an issue if the test suite outputted
files that differ from the stored files causing git to complain about
modified files. Since they are to be ignored git doesn't need to track
them.